+++ include/sdf/SDFImpl.hh
@@ -94,6 +94,9 @@ namespace sdf
     /// \brief Set SDF values from a string
     public: void SetFromString(const std::string &_sdfData);
 
+    /// \brief Clear the data in this object.
+    public: void Clear();
+
     /// \brief Get a pointer to the root element
     /// \return Pointer to the root element
     public: ElementPtr Root() const;
@@ -109,6 +112,14 @@ namespace sdf
     /// \brief Set the path on disk.
     /// \param[in] _path Path on disk.
     public: void SetFilePath(const std::string &_path);
+
+    /// \brief Set the spec version that this was originally parsed from.
+    /// \param[in] _version Spec version string.
+    public: void SetOriginalVersion(const std::string &_version);
+
+    /// \brief Get the spec version that this was originally parsed from.
+    /// \return Spec version string.
+    public: const std::string &OriginalVersion() const;
 
     /// \brief Get the version
     /// \return The version as a string