+++ src/SDFImplPrivate.hh
@@ -43,6 +43,9 @@ namespace sdf
 
     /// \brief Path to SDF document on disk.
     public: std::string path;
+
+    /// \brief Spec version that this was originally parsed from.
+    public: std::string originalVersion;
   };
   /// \}
 }